Replace continous mlp baseline w model primitive
- Removed continuous mlp baseline, and rename the model
  primitive version as just continuous mlp baseline.
- Replaced cmb_with_model with cmb in places where mentioned
  in the codebase.
