+++ src/garage/tf/regressors/__init__.py
@@ -11,8 +11,6 @@ from garage.tf.regressors.categorical_mlp_regressor_with_model import (
     CategoricalMLPRegressorWithModel)
 from garage.tf.regressors.continuous_mlp_regressor import (
     ContinuousMLPRegressor)
-from garage.tf.regressors.continuous_mlp_regressor_with_model import (
-    ContinuousMLPRegressorWithModel)
 from garage.tf.regressors.gaussian_cnn_regressor_model import (
     GaussianCNNRegressorModel)
 from garage.tf.regressors.gaussian_cnn_regressor_with_model import (
@@ -23,8 +21,8 @@ from garage.tf.regressors.gaussian_mlp_regressor import GaussianMLPRegressor
 __all__ = [
     'BernoulliMLPRegressor', 'BernoulliMLPRegressorWithModel',
     'CategoricalMLPRegressor', 'CategoricalMLPRegressorWithModel',
-    'ContinuousMLPRegressor', 'ContinuousMLPRegressorWithModel',
-    'GaussianCNNRegressorModel', 'GaussianCNNRegressorWithModel',
-    'GaussianMLPRegressor', 'GaussianConvRegressor', 'Regressor2',
-    'StochasticRegressor', 'StochasticRegressor2'
+    'ContinuousMLPRegressor', 'GaussianCNNRegressorModel',
+    'GaussianCNNRegressorWithModel', 'GaussianMLPRegressor',
+    'GaussianConvRegressor', 'Regressor2', 'StochasticRegressor',
+    'StochasticRegressor2'
 ]
