+++ tests/garage/tf/baselines/test_continuous_mlp_baseline.py
@@ -5,22 +5,23 @@ import numpy as np
 import pytest
 import tensorflow as tf
 
-from garage.tf.baselines import ContinuousMLPBaselineWithModel
+from garage.tf.baselines import ContinuousMLPBaseline
 from garage.tf.envs import TfEnv
 from tests.fixtures import TfGraphTestCase
 from tests.fixtures.envs.dummy import DummyBoxEnv
 from tests.fixtures.regressors import SimpleMLPRegressor
 
 
-class TestContinuousMLPBaselineWithModel(TfGraphTestCase):
+class TestContinuousMLPBaseline(TfGraphTestCase):
+
     @pytest.mark.parametrize('obs_dim', [[1], [2], [1, 1], [2, 2]])
     def test_fit(self, obs_dim):
         box_env = TfEnv(DummyBoxEnv(obs_dim=obs_dim))
         with mock.patch(('garage.tf.baselines.'
-                         'continuous_mlp_baseline_with_model.'
-                         'ContinuousMLPRegressorWithModel'),
+                         'continuous_mlp_baseline.'
+                         'ContinuousMLPRegressor'),
                         new=SimpleMLPRegressor):
-            cmb = ContinuousMLPBaselineWithModel(env_spec=box_env.spec)
+            cmb = ContinuousMLPBaseline(env_spec=box_env.spec)
         paths = [{
             'observations': [np.full(obs_dim, 1)],
             'returns': [1]
@@ -38,16 +39,15 @@ class TestContinuousMLPBaselineWithModel(TfGraphTestCase):
     def test_param_values(self, obs_dim):
         box_env = TfEnv(DummyBoxEnv(obs_dim=obs_dim))
         with mock.patch(('garage.tf.baselines.'
-                         'continuous_mlp_baseline_with_model.'
-                         'ContinuousMLPRegressorWithModel'),
+                         'continuous_mlp_baseline.'
+                         'ContinuousMLPRegressor'),
                         new=SimpleMLPRegressor):
-            cmb = ContinuousMLPBaselineWithModel(env_spec=box_env.spec)
-            new_cmb = ContinuousMLPBaselineWithModel(
-                env_spec=box_env.spec, name='ContinuousMLPBaselineWithModel2')
+            cmb = ContinuousMLPBaseline(env_spec=box_env.spec)
+            new_cmb = ContinuousMLPBaseline(env_spec=box_env.spec,
+                                            name='ContinuousMLPBaseline2')
 
-        # Manual change the parameter of ContinuousMLPBaselineWithModel
-        with tf.compat.v1.variable_scope(
-                'ContinuousMLPBaselineWithModel2', reuse=True):
+        # Manual change the parameter of ContinuousMLPBaseline
+        with tf.compat.v1.variable_scope('ContinuousMLPBaseline2', reuse=True):
             return_var = tf.compat.v1.get_variable('SimpleMLPModel/return_var')
         return_var.load(1.0)
 
@@ -62,26 +62,25 @@ class TestContinuousMLPBaselineWithModel(TfGraphTestCase):
     def test_get_params_internal(self, obs_dim):
         box_env = TfEnv(DummyBoxEnv(obs_dim=obs_dim))
         with mock.patch(('garage.tf.baselines.'
-                         'continuous_mlp_baseline_with_model.'
-                         'ContinuousMLPRegressorWithModel'),
+                         'continuous_mlp_baseline.'
+                         'ContinuousMLPRegressor'),
                         new=SimpleMLPRegressor):
-            cmb = ContinuousMLPBaselineWithModel(env_spec=box_env.spec)
+            cmb = ContinuousMLPBaseline(env_spec=box_env.spec)
         params_interal = cmb.get_params_internal()
         trainable_params = tf.compat.v1.trainable_variables(
-            scope='ContinuousMLPBaselineWithModel')
+            scope='ContinuousMLPBaseline')
         assert np.array_equal(params_interal, trainable_params)
 
     def test_is_pickleable(self):
         box_env = TfEnv(DummyBoxEnv(obs_dim=(1, )))
         with mock.patch(('garage.tf.baselines.'
-                         'continuous_mlp_baseline_with_model.'
-                         'ContinuousMLPRegressorWithModel'),
+                         'continuous_mlp_baseline.'
+                         'ContinuousMLPRegressor'),
                         new=SimpleMLPRegressor):
-            cmb = ContinuousMLPBaselineWithModel(env_spec=box_env.spec)
+            cmb = ContinuousMLPBaseline(env_spec=box_env.spec)
         obs = {'observations': [np.full(1, 1), np.full(1, 1)]}
 
-        with tf.compat.v1.variable_scope(
-                'ContinuousMLPBaselineWithModel', reuse=True):
+        with tf.compat.v1.variable_scope('ContinuousMLPBaseline', reuse=True):
             return_var = tf.compat.v1.get_variable('SimpleMLPModel/return_var')
         return_var.load(1.0)
